Refactor CRC32_MATCH impl package
